Refactor schema interfaces and move types to node
Moved IDB and INodeDB interfaces from chainkit/schema/types.go to node/schema/idb.go as IDBChainkit and IDBTool, respectively. Deleted types.go and updated references in chainkit.go. Renamed config.go to schema.go and added DownloadResult type to schema.go for better organization.
